+++ Sources/AnimalKingdomAPI/AnimalKingdomAPI/Sources/Operations/ClassroomPetsQuery.swift
@@ -140,7 +140,7 @@ public class ClassroomPetsQuery: GraphQLQuery {
         public var species: String { data["species"] }
         public var humanName: String? { data["humanName"] }
         public var laysEggs: Bool { data["laysEggs"] }
-        public var wingspan: Float { data["wingspan"] }
+        public var wingspan: Double { data["wingspan"] }
 
         public struct Fragments: FragmentContainer {
           public let data: DataDict
